requirements.txt
Beaker==1.6.4
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
Mako==1.0.0
Markdown==2.5.1
MarkupSafe==0.23
Pillow==2.6.0
Werkzeug==0.9.6
gunicorn==19.1.1
#httplib2==0.9
image==1.3.3
itsdangerous==0.24
#lxml==3.4.0
#python-distutils-extra==2.38
SQLAlchemy==0.9.8
sqlalchemy-migrate==0.9.2
wsgiref==0.1.2
#wxPython==3.0.1.1
#wxPython-common==3.0.1.1
